tiny cleanup, buttons super cursed :<
