+++ Cargo.toml
@@ -46,7 +46,7 @@ wayland-client = { version = "0.31.1", optional = true }
 wayland-protocols = { version = "0.31.0", features = ["client", "unstable"], optional = true }
 
 # Windows Ink `RealTimeStylus`
-[target.'cfg(any(target_os = "windows"))'.dependencies.windows]
+[target.'cfg(any(target_os = "windows", unix))'.dependencies.windows]
 version = "0.52.0"
 optional = true
 features = [
