+++ src/lib.rs
@@ -51,31 +51,6 @@ pub mod util;
 pub use builder::Builder;
 use events::Events;
 
-pub(crate) mod macro_bits {
-    /// Implements an public opaque ID,
-    /// assuming the struct has a `internal_id` which implements `Into<platform::InternalID>`
-    macro_rules! impl_get_id {
-        ($id_name:ident for $impl_for:ident) => {
-            /// An opaque ID. Can be used to keep track of hardware, but only during its lifetime.
-            /// Once the hardware is `Removed`, the ID loses meaning.
-            #[derive(Clone, Debug, Hash, PartialEq, Eq)]
-            #[allow(clippy::module_name_repetitions)]
-            pub struct $id_name(crate::platform::InternalID);
-
-            impl $impl_for {
-                /// Opaque, transient ID of this tool, assigned arbitrarily by the software. Will not
-                /// be stable across invocations or even unplugs/replugs!
-                #[must_use]
-                pub fn id(&self) -> $id_name {
-                    $id_name(self.internal_id.clone().into())
-                }
-            }
-        };
-    }
-    // Weird hacks to allow use from submodules..
-    pub(crate) use impl_get_id;
-}
-
 /// A trait that every object is.
 /// Used to cast things to `dyn Erased` which leaves us with a wholly erased type.
 trait Erased {}
